limited_functionality/docker/requirements.txt
pyaudio
vosk
mysql-connector-python
python-dotenv